% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{add_delete_column}
\alias{add_delete_column}
\title{add delete column}
\usage{
add_delete_column(df, ns, ...)
}
\arguments{
\item{df}{original tibble to add a delete column to}

\item{ns}{shiny namespace object}

\item{...}{currently unused, additional arguments passed on to internal methods}
}
\value{
a tibble
}
\description{
Add a column with a delete button to a tibble, attaching a namespaced shiny input event
for listening to the click
}
