% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{add_show_column}
\alias{add_show_column}
\title{add show column}
\usage{
add_show_column(df, ns, ...)
}
\arguments{
\item{df}{original tibble to add a show column to}

\item{ns}{shiny namespace object}

\item{...}{currently unused, additional arguments passed on to internal methods}
}
\value{
a tibble
}
\description{
Add a column with a show button to a tibble, attaching a namespaced shiny input event
for listening to the click
}
