% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tabmodule.R
\name{advanced_tab_module_server}
\alias{advanced_tab_module_server}
\title{advanced_tab module server}
\usage{
advanced_tab_module_server(input, output, session, data, model, scoring_funcs)
}
\arguments{
\item{input}{necessary input arg for shiny server function}

\item{output}{necessary output arg for shiny server function}

\item{session}{necessary session arg for shiny server function}
}
\description{
advanced_tab module server
}
