% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tabmodule.R
\name{advanced_tab_module_ui}
\alias{advanced_tab_module_ui}
\title{advanced_tab module ui}
\usage{
advanced_tab_module_ui(id)
}
\arguments{
\item{id}{shiny id necessary for module}
}
\description{
advanced_tab module ui
}
