% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{app_ui}
\alias{app_ui}
\title{app_ui}
\usage{
app_ui(req, question_data, default_response)
}
\arguments{
\item{req}{Internal parameter for \code{{shiny}}.
DO NOT REMOVE.}

\item{question_data}{The question data, a list, likely parsed from a YAML file that specifies the questions to be
displayed in the application}

\item{default_response}{The default values, used to initialise the model}
}
\description{
Defines the user interface for the DIAGRAM application. This content is a heavily revised
version of a script built for DiAGRAM by the University of Warwick and The National
Archive. Stephen James Krol, Monash University, Melbourne, stephen.james.krol@gmail.com
}
