% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bind_questions}
\alias{bind_questions}
\title{bind questions}
\usage{
bind_questions(questions, responses)
}
\arguments{
\item{questions}{A list of all the question data}

\item{responses}{A named list of responses}
}
\description{
Will recursively bind questions and their sub elements to the reponses
}
