% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_test.R
\name{body_to_gateway_payload}
\alias{body_to_gateway_payload}
\title{Inserts a request body into AWS' API Gateway's payload format}
\usage{
body_to_gateway_payload(body)
}
\arguments{
\item{body}{String representing request body}
}
\description{
Inserts a request body into AWS' API Gateway's payload format
}
