% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{build_csv_data}
\alias{build_csv_data}
\title{Prepare data for CSV report from responses}
\usage{
build_csv_data(responses)
}
\arguments{
\item{responses}{a set of responses (tibble) of class simple_responses}
}
\description{
Prepare data for CSV report from responses
}
