% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calculate_utility}
\alias{calculate_utility}
\title{Calculate utility}
\usage{
calculate_utility(model)
}
\arguments{
\item{model}{a model of class bn.fit}
}
\value{
A named list of two scores
}
\description{
Calculate the utility metrics Intellectual Control and
Renderability from the given bayesian network model
}
