% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check}
\alias{check}
\title{Generic functions for checking assertions on response nodes.}
\usage{
check(x, ...)
}
\arguments{
\item{x:}{typically single response node, may be a
numeric or character vector, or a list of such objects.}

\item{options}{named vector (or list) of allowed options for
a response}
}
\description{
Generic functions for checking assertions on response nodes.
}
