% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart.R
\name{column_chart}
\alias{column_chart}
\title{Column chart}
\usage{
column_chart(data, xlabel)
}
\arguments{
\item{data}{data to be displayed}

\item{xlabel}{character, x axis label for the plot}
}
\description{
Creates the standard column chart used in many tabs with set colours and themeing
}
