% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{create_data_modal}
\alias{create_data_modal}
\title{create data modal}
\usage{
create_data_modal(ns, data)
}
\arguments{
\item{ns}{a shiny namespace object}

\item{data}{The data that will be shown, only used to determine whether this
is a custom model or not as at this time custom model responses can not be shown
sensibly}
}
\value{
a shiny taglist object
}
\description{
Create the modal object that contains the shown responses for
a selected model
}
