% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{createModal}
\alias{createModal}
\title{createModal}
\usage{
createModal(ns, model, policy, comment)
}
\arguments{
\item{ns}{a shiny namespace object}

\item{model}{the model name of the selected model}

\item{policy}{the scenario name of the selected scenario}

\item{comment}{the current comment string that is attributed to the
selected model/scenario}
}
\value{
a shiny taglist object
}
\description{
Create the modal dialogue box for editing names and comments of models/scenarios
}
