% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_about_tab.R
\name{definitions_tab}
\alias{definitions_tab}
\title{learn about/definitions tab}
\usage{
definitions_tab()
}
\value{
a shiny tag list object
}
\description{
Creates the learn about page from the learn_about.md file in the inst/text_content file
}
