% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserialise.R
\name{deserialise}
\alias{deserialise}
\title{This function transforms a request as received by API Gateway into the
format used by the earlier {diagramAPI} package.}
\usage{
deserialise(event_content)
}
\arguments{
\item{event_content}{Raw payload as received from API Gateway. See the \href{AWS Documentation}{https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html#http-api-develop-integrations-lambda.proxy-format} for more details on the format of this payload.}
}
\description{
Adhering to the format used by {diagramAPI} allows reuse of all of
{diagramAPI}'s deserialisation routines and json wrangling.
}
