% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_bannermodule.R
\name{dev_banner_module_server}
\alias{dev_banner_module_server}
\title{dev_banner module server}
\usage{
dev_banner_module_server(input, output, session)
}
\arguments{
\item{input}{necessary input arg for shiny server function}

\item{output}{necessary output arg for shiny server function}

\item{session}{necessary session arg for shiny server function}
}
\description{
dev_banner module server
}
