% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_bannermodule.R
\name{dev_banner_module_ui}
\alias{dev_banner_module_ui}
\title{dev_banner module ui}
\usage{
dev_banner_module_ui(id)
}
\arguments{
\item{id}{shiny id necessary for module}
}
\description{
dev_banner module ui
}
