% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{diff_table}
\alias{diff_table}
\title{Build a table of diffs}
\usage{
diff_table(obj)
}
\arguments{
\item{obj}{a collection of models/policies (full responses),
each with the same $name}
}
\description{
A table which builds a table showing only
the responses to questions that are different
to the base model.
}
