% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\docType{data}
\name{.model_prob_names}
\alias{.model_prob_names}
\title{Model node to probability table names map}
\format{
An object of class \code{list} of length 9.
}
\usage{
.model_prob_names
}
\description{
each node in the network has fixed outcomes which are attributed
probabilities. This object allows a map between the node names
and the names of the probability tables in the network object.
}
\keyword{datasets}
