% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ensure_grain_model}
\alias{ensure_grain_model}
\title{Ensure grain model}
\usage{
ensure_grain_model(model)
}
\arguments{
\item{model}{model object, likely of class bn.fit}
}
\value{
model object of class grain
}
\description{
Ensure that model has type grain required for
calculating score using the gRain library.
}
