% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{extract_scores}
\alias{extract_scores}
\title{Extract scores from data}
\usage{
extract_scores(parsed_json)
}
\arguments{
\item{parsed_json}{raw data as parsed from json (list)}
}
\description{
Extract scores from data
}
