% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{find_diff}
\alias{find_diff}
\title{Find diff}
\usage{
find_diff(download_data, model_name)
}
\arguments{
\item{download_data}{Data formatted ready for download see \code{format_data_for_download}}

\item{model_name}{Character, model name for which to compare responses on policies}
}
\value{
a named list of the policy response differences
}
\description{
Find the change in responses for all policies for a given model
}
