% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{find_scenario_diffs}
\alias{find_scenario_diffs}
\title{Calculate model/scenario diff}
\usage{
find_scenario_diffs(responses)
}
\arguments{
\item{responses, }{a whole group of responses that come from the same model name}
}
\description{
Calculate the difference between each scenario and the base-model
in a grouped set of responses with the same base-model $name.
Exported for pdf.
}
