% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{format_data_for_download}
\alias{format_data_for_download}
\title{format data for download}
\usage{
format_data_for_download(data, question_data)
}
\arguments{
\item{data}{a tibble of model/scenario response and comment data}

\item{question_data}{A named list of questions}
}
\value{
a tibble
}
\description{
Will format data such that the questions and responses are kept readable together
}
