% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{format_model_table}
\alias{format_model_table}
\title{format model table}
\usage{
format_model_table(intermediate, model, scoring_funcs, show_policy)
}
\arguments{
\item{intermediate}{an intermediate set of unformatted model/scenario data (tibble)}

\item{model}{the underlying bayesian network model to produce the scores, the actual values of the probabilities
internal to this model should in theory be irrelevant as they will be replaced with the values
generated by the set of responses for a given model/scenario set by the user}

\item{scoring_funcs}{a named list of scoring functions to apply to the responses in order to convert
them to probabilities}

\item{show_policy}{logical, should the policies be included in the formatted data}
}
\value{
a tibble with models, comments and scores
}
\description{
Ensures formatting across model table displays is consistent and calculates each models scores
}
