% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{format_responses}
\alias{format_responses}
\title{Format responses}
\usage{
format_responses(values)
}
\arguments{
\item{values}{named list of question responses}
}
\value{
a tibble(node, response)
}
\description{
Take list of question responses and restructure to nice tibble
}
