% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{formulate_question}
\alias{formulate_question}
\title{formulate question}
\usage{
formulate_question(question, default_response, ns)
}
\arguments{
\item{question}{a list that describes the content of the question}

\item{default_response}{a named list component that contains the initialising value for the question}

\item{ns}{a namespacing function, typically as returned by shiny::NS}
}
\value{
a named list of question components
}
\description{
This function will parse the question data stored within a list, match the question type
against the appropriate module functions, create the user interface element in it's default
state and return a list of components.
}
