% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_api.R
\name{create_routes}
\alias{create_routes}
\alias{generate_api}
\title{create_routes: For a given directory, create a named vector of files and endpoints}
\usage{
create_routes(dir)

generate_api(routes, ...)
}
\arguments{
\item{dir}{directory containing end points}

\item{routes}{A list of file. Element names are the plumber endpoints}

\item{...}{filters and/or envir args to pass to plumber::pr}
}
\value{
A Plumber object
}
\description{
routes is typically created by create_routes
}
