% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_base_model}
\alias{get_base_model}
\title{Get the base model}
\usage{
get_base_model(responses)
}
\arguments{
\item{responses, }{a whole group of responses}
}
\description{
From a list of responses, return the one which represents a base model
}
