% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_comments}
\alias{get_comments}
\title{Get comment data}
\usage{
get_comments(obj)
}
\arguments{
\item{obj}{a full collection of models/policies (responses)}
}
\description{
Convenience function exported for pdf template
to grab the scenario names from a set of models/policies
(responses)
}
