% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_option_val.R
\name{get_option_val}
\alias{get_option_val}
\title{Get option val}
\usage{
get_option_val(res)
}
\arguments{
\item{res}{typically single response node, may be a
numeric or character vector, or a list of such objects.}
}
\description{
Generic function for getting allowed options for
single node responses.
}
