% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_questions_tibble}
\alias{get_questions_tibble}
\title{Get a formatted set of questions}
\usage{
get_questions_tibble()
}
\description{
Convenience function exported for building pdfs that
grabs the question and answer text used in the front end
and formats it as tibble.
}
