% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{group_objects_by_model}
\alias{group_objects_by_model}
\title{Group similar models}
\usage{
group_objects_by_model(obj)
}
\arguments{
\item{obj}{a full collection of models/policies (responses)}
}
\description{
Group sets of responses by their base-model $name
}
