% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{handler}
\alias{handler}
\title{Route requests from API Gateway}
\usage{
handler(event_content, context)
}
\arguments{
\item{event_content}{Raw payload as received from API Gateway. See the \href{AWS Documentation}{https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html#http-api-develop-integrations-lambda.proxy-format} for more details on the format of this payload.}

\item{context}{Metadata about the received event}
}
\description{
There are a few different approaches to emulating multiple endpoints with
AWS Lamba and AWS API Gateway.
\href{This StackOverflow question}{https://stackoverflow.com/questions/41425511/aws-api-gateway-lambda-multiple-endpoint-functions-vs-single-endpoint}
provides a good overview of some of the pros and cons of the different
possible approaches.
}
