% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_tab.R
\name{home_tab}
\alias{home_tab}
\title{home tab}
\usage{
home_tab()
}
\value{
a shiny taglist object
}
\description{
Function generates the home page/landing oage for the diagram app, grabbing assets from the inst
file from the package.
}
