% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{join_questions_responses}
\alias{join_questions_responses}
\title{Bind questions to repsonses}
\usage{
join_questions_responses(questions, responses, ...)
}
\arguments{
\item{questions}{a tibble of questions, see \code{get_questions_tibble()}}

\item{responses}{a tibble of responses for a model/scenario}

\item{...}{arguments passed to method, currently unused}
}
\description{
Generic (for simple_responses and scenario_diff) for
building a single tibble that contains nicely formatted
questions and responses for the pdfs.
}
