% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_default_model}
\alias{load_default_model}
\title{Load default model}
\usage{
load_default_model()
}
\value{
object of class bn.fit
}
\description{
Load the default .bif model that should seed model scoring
This file was generated by the original authors.
}
