% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{load_single_response}
\alias{load_single_response}
\title{Load single response}
\usage{
load_single_response(path)
}
\value{
A named list of the form \code{(<node_name> = <response>)}
}
\description{
Load a single set of model question responses from disk
}
