% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_report.R
\name{make_csv_report}
\alias{make_csv_report}
\title{Generate CSV report from request object}
\usage{
make_csv_report(req)
}
\arguments{
\item{req}{A request object satisfying the \href{Rook interface}{https://cran.r-project.org/web/packages/Rook/README.html} between web servers and R applications.}
}
\description{
Generate CSV report from request object
}
