% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{model_table_module_ui}
\alias{model_table_module_ui}
\title{model table module ui}
\usage{
model_table_module_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module required by all shiny modules}
}
\value{
a shiny taglist object
}
\description{
module ui container for the table that displays model information
}
