% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_probability.R
\name{numeric_to_probability}
\alias{numeric_to_probability}
\title{Numeric to Probability}
\usage{
numeric_to_probability(numeric_response)
}
\arguments{
\item{numeric_response}{a named list of numeric values}
}
\description{
Convert a list of numeric responses to model form
questions into approriate probability tables for
executing the model.
}
