% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_probability.R
\name{numeric_to_probability_single}
\alias{numeric_to_probability_single}
\title{Numeric to Probability}
\usage{
numeric_to_probability_single(res, name)
}
\arguments{
\item{res}{numeric vector}

\item{name}{character, name of corresponding node in network}
}
\description{
Convert a single numeric response from the model form
answers to a probability.
}
