\name{openplotmat}
\alias{openplotmat}
\title{
  Creates an empty plot used for diagram plotting.
}
\description{
  Creates a plotting region, bounded by [0,1]
  without axes, labels, titles
}
\usage{
openplotmat (asp = NA, ...)
}
\arguments{
  \item{asp }{the y/x aspect ratio.
  }
  \item{... }{arguments passed to \code{emptyplot} from package \code{shape}.
  }
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\seealso{
  \code{emptyplot} from package \code{shape}.
}
\keyword{hplot}

