% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_module.R
\name{parse_column_event}
\alias{parse_column_event}
\title{parse column event}
\usage{
parse_column_event(idstr)
}
\arguments{
\item{idstr}{An id string, typically sent with a shiny click event}
}
\value{
an integer corresponding to which button has been clicked
}
\description{
A function to help deal with the many edit, show and delete buttons added
into the tables of the model view. Parses the string and matches the corresponding number.
}
