% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report_style.R
\name{pdf_report}
\alias{pdf_report}
\title{Custom format for NATA reports, based on \code{bookdown::pdf_book}}
\usage{
pdf_report(
  keep_tex = TRUE,
  latex_engine = "xelatex",
  toc = FALSE,
  toc_depth = 2,
  number_sections = FALSE,
  extra_pandoc_args = NULL,
  ...
)
}
\arguments{
\item{keep_tex}{Whether or not to keep .tex file. Default \code{FALSE}}

\item{latex_engine}{Engine used to compile pdf. Default \code{'pdflatex'}}

\item{toc}{Whether or not to show table of contents. Default \code{FALSE}}

\item{toc_depth}{Depth of levels to show in toc \code{2}}

\item{number_sections}{Whether or not to number chapters, sections, etc. Default \code{FALSE}}

\item{extra_pandoc_args}{Extra arguments for pandoc e.g
\code{c(
  rmarkdown::pandoc_variable_arg("logo_img", "example.png"),
  rmarkdown::pandoc_metadata_arg("title","example")
)}}

\item{...}{Passed on to \code{bookdown::pdf_book(...)}}
}
\description{
Custom format for NATA reports, based on \code{bookdown::pdf_book}
}
