% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_table_part}
\alias{pdf_table_part}
\title{Build a single row of the table for pdf}
\usage{
pdf_table_part(question, response, node)
}
\arguments{
\item{question}{either a list or a character vector}

\item{response}{either a list of equivalent shape to question, or a vector}

\item{node}{character label of node corresponding to question}
}
\description{
Collapses questions and responses with multiple parts
and structure to a single consistent tibble chunk
}
