% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_plot.R
\name{plot}
\alias{plot}
\title{Generate plot from request object}
\usage{
plot(req)
}
\arguments{
\item{req}{A request object satisfying the \href{Rook interface}{https://cran.r-project.org/web/packages/Rook/README.html} between web servers and R applications.}
}
\description{
Generate plot from request object
}
