% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_visualisationmodule.R
\name{policy_bar_chart}
\alias{policy_bar_chart}
\title{Create bar chart for visualisation module}
\usage{
policy_bar_chart(policy_data)
}
\arguments{
\item{policy_data}{Model and policy data. See inst/table_experiment.R for temp data}
}
\description{
Create bar chart for visualisation module
}
