% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_visualisationmodule.R
\name{policy_bar_gg}
\alias{policy_bar_gg}
\title{policy bar gg}
\usage{
policy_bar_gg(policy_data)
}
\arguments{
\item{policy_data}{a tibble of model/scenario data}
}
\description{
Creates a nice bar chart using ggplot with fixed colours for policies and
models
}
