% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_creation.R
\name{policy_creation_module_server}
\alias{policy_creation_module_server}
\title{policy creation module server}
\usage{
policy_creation_module_server(
  input,
  output,
  session,
  input_data,
  question_data,
  model,
  scoring_funcs
)
}
\arguments{
\item{input}{necessary for shiny module}

\item{output}{necessary for shiny module}

\item{session}{necessary for shiny module}

\item{input_data}{The data which lays out the list of existing models/scenarios}

\item{question_data}{The named list of questions}

\item{model}{The underlying bayesian network model}

\item{scoring_funcs}{A named list of functions which translate user inputs into probabilities for the mode}
}
\description{
Server side logic to handle policy creation
}
