% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_creation.R
\name{policy_creation_module_ui}
\alias{policy_creation_module_ui}
\title{policy creation module ui}
\usage{
policy_creation_module_ui(id)
}
\arguments{
\item{id}{namespace id requried by shiny modules}
}
\description{
Lays out the user interface for the scenario building, choosing the model
to build a policy from, the questions and then cycling through the questions to
build a set of responses
}
