% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_visualisationmodule.R
\name{policy_visualisation_module_server}
\alias{policy_visualisation_module_server}
\title{policy_visualisation module server}
\usage{
policy_visualisation_module_server(
  input,
  output,
  session,
  model_data,
  model,
  scoring_funcs,
  question_data
)
}
\arguments{
\item{input}{necessary input arg for shiny server function}

\item{output}{necessary output arg for shiny server function}

\item{session}{necessary session arg for shiny server function}
}
\description{
server side logic for the visualisation of model/scenario data
}
