% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_visualisationmodule.R
\name{policy_visualisation_module_ui}
\alias{policy_visualisation_module_ui}
\title{policy_visualisation module ui}
\usage{
policy_visualisation_module_ui(id)
}
\arguments{
\item{id}{shiny id necessary for module}
}
\description{
policy_visualisation module ui
}
