% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_bar_chart_data}
\alias{prepare_bar_chart_data}
\title{Prepare data to be plotted as a bar chart}
\usage{
prepare_bar_chart_data(data)
}
\arguments{
\item{data}{raw list data as parsed from json}
}
\description{
Prepare data to be plotted as a bar chart
}
