% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{prepare_csv}
\alias{prepare_csv}
\title{Prepare CSV}
\usage{
prepare_csv(data, file = tempfile(fileext = ".csv"))
}
\arguments{
\item{data}{Tibble to be stored}

\item{file}{A file location to write to}
}
\value{
The file location
}
\description{
Will prepare the data for CSV download
}
