% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{prepare_json}
\alias{prepare_json}
\title{Prepare json}
\usage{
prepare_json(data, file = tempfile(fileext = ".json"))
}
\arguments{
\item{data}{Tibble to be stored}

\item{file}{A file location to write to}
}
\value{
The file location
}
\description{
Will prepare the data in a JSON format which is suitable for upload at a later date
}
