% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{prepare_pdf}
\alias{prepare_pdf}
\title{prepare pdf}
\usage{
prepare_pdf(
  data,
  question_data,
  model,
  scoring_funcs,
  template = system.file("assets", "templates", "pdf_template.Rmd", package =
    "diagramNAT"),
  file = tempfile(fileext = ".pdf")
)
}
\arguments{
\item{data}{a tibble containing the data for models/scenarios}

\item{question_data}{A named list of question data}

\item{model}{The underlying bayesian network model}

\item{scoring_funcs}{A named list of functions which translate user responses into probabilities}

\item{template}{A pdf template to use for rendering the output}

\item{file}{A file path to act as a target for the rendered pdf prior to download.}
}
\value{
a character, file path to the rendered pdf
}
\description{
Create a downloadable pdf from the selected responses to the models/scenarios
}
