% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questions_module_server}
\alias{questions_module_server}
\title{questions module server}
\usage{
questions_module_server(
  input,
  output,
  session,
  question_data,
  default_response,
  is_policy = FALSE
)
}
\arguments{
\item{input}{necessary input arg for shiny server function}

\item{output}{necessary output arg for shiny server function}

\item{session}{necessary session arg for shiny server function}

\item{question_data}{A named list of question data, the same as is used to build the UI elements}

\item{default_response}{A named list of default responses to the questions}

\item{is_policy}{logical, whether or not this module is being used for scenario creation}
}
\value{
A named list of reactive objects to pass back to the parent, see details
}
\description{
Manage the server side logic for the processing of questions
}
\details{
Returned list of reactive objects
list(state = return_val, name = reactive(input$name), comments = reactive(input$comment), finish = reactive(input$finish),
\itemize{
\item state - the current response state of the questions
\item name - the name given to the model or scenario
\item comments - the user input comments supplied alongside the responses
\item finish - the click event for finishing the model building process
\item go - the click event for pressing the next button
}
}
