% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questions_module_ui}
\alias{questions_module_ui}
\title{questions module ui}
\usage{
questions_module_ui(id, question_data, default_response, is_policy = FALSE)
}
\arguments{
\item{id}{A unique identifier for the module, required by all shiny modules}
}
\value{
A shiny taglist object
}
\description{
Main module user interface for the combination of questions. All UI elements are built and rendered
on initialisation, sequencing through the questions is done by showing and hiding the sequence of
individual UI elements.
}
