% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tabmodule.R
\name{report_tab_module_server}
\alias{report_tab_module_server}
\title{report_tab module server}
\usage{
report_tab_module_server(
  input,
  output,
  session,
  data,
  model,
  question_data,
  scoring_funcs
)
}
\arguments{
\item{input}{necessary input arg for shiny server function}

\item{output}{necessary output arg for shiny server function}

\item{session}{necessary session arg for shiny server function}

\item{data}{reactive data source for table}

\item{model}{The bnlearn model to calculate scores}

\item{scoring_funcs}{a named list of custom scoring functions to apply to raw responses}
}
\description{
report_tab module server
}
