% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tabmodule.R
\name{report_tab_module_ui}
\alias{report_tab_module_ui}
\title{report_tab module ui}
\usage{
report_tab_module_ui(id)
}
\arguments{
\item{id}{shiny id necessary for module}
}
\description{
report_tab module ui
}
