% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the diagram app}
\usage{
run_app(n = 9)
}
\arguments{
\item{n}{Number of questions, defaults to 9}
}
\description{
Runs the diagram app using the partial function from purrr.
}
