% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_model.R
\name{score_model}
\alias{score_model}
\title{Score model}
\usage{
score_model(req, model = load_default_model())
}
\arguments{
\item{req}{a request object satisfying the Rook interface.}

\item{model}{a network model, likely of class bn.fit}
}
\value{
list with the two scores and the nodes information
}
\description{
Currently method stub, used to obtain the renderability and
intellectual control scores for a given set of model inputs.
}
