% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{score_model_}
\alias{score_model_}
\title{Score model}
\usage{
score_model_(responses, model)
}
\arguments{
\item{responses}{a parsed set of responses}

\item{model}{a network model, likely of class bn.fit}
}
\description{
Main entry point for scoring a model to be called by
the exposed API /score_model function. Should function
appropriately for both simple and advanced responses
(as indicated by the class attribute of responses parameter).
Will call necessary preprocessing on the responses and model
before returning a score.
}
