% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_serialisers.R
\name{serialise_binary_data}
\alias{serialise_binary_data}
\title{Serialise binary data to pass as response to API Gateway}
\usage{
serialise_binary_data(
  body,
  content_type = c("application/pdf", "image/png"),
  status_code = 200L,
  headers = list()
)
}
\arguments{
\item{body}{The content to be serialised}

\item{content_type}{The type of binary content to be returned}

\item{status_code}{The HTTP response status code to return}

\item{headers}{Additional content headers to append}
}
\description{
Serialise binary data to pass as response to API Gateway
}
