% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_serialisers.R
\name{serialise_r}
\alias{serialise_r}
\title{Serialise R object to pass as response to API Gateway}
\usage{
serialise_r(body, status_code = 200L, headers = list())
}
\arguments{
\item{body}{The content to be serialised}

\item{status_code}{The HTTP response status code to return}

\item{headers}{Additional content headers to append}
}
\description{
Serialise R object to pass as response to API Gateway
}
\details{
Responses posted from AWS Lambda to AWS API Gateway must adhere to
the format outlined at: https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html#http-api-develop-integrations-lambda.response
}
