% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_test.R
\name{test_data_to_gateway_payload}
\alias{test_data_to_gateway_payload}
\title{Requests coming via AWS' API Gateway adhere to the payload format outlined at:
https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html}
\usage{
test_data_to_gateway_payload(
  test_dir = c("csv_requests", "error_generators", "pdf_requests", "plot_requests",
    "score_requests"),
  test_file = list.files(system.file("extdata", "test_data", test_dir, package =
    "diagramLambda"))
)
}
\description{
This function helps convert the tests in extdata/test_data into the format
which they would be recieved by a Lambda function sitting behind API Gateway
}
