% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_req_score}
\alias{test_req_score}
\title{Test Req}
\usage{
test_req_score(
  type = c("simple", "advanced", "simple_scenario", "mixed", "mixed_simple",
    "simple_single")
)
}
\arguments{
\item{type}{character reflecting which type of responses to generate.}
}
\value{
an environment mimicking the necessary parts of Rook request
object that are required for testing purposes.
}
\description{
Utility for creating fake requests to the model_score endpoint.
Used for testing functions related to processing model responses
and generating scores.
}
