% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_numeric.R
\name{to_numeric}
\alias{to_numeric}
\title{To numeric}
\usage{
to_numeric(res)
}
\arguments{
\item{res}{response (typically numeric or character vector) for a single node}
}
\value{
a numeric vector
}
\description{
Generic function for calling to_numeric on response nodes.
First calls the generic check() for assertion checking the recieved
response value.
}
