% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_node}
\alias{validate_node}
\title{Validate data on a single node}
\usage{
validate_node(res)
}
\arguments{
\item{res}{result object for a single node in the bayesian
network}
}
\description{
Validate data on a single node
}
