% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{write_temp_png}
\alias{write_temp_png}
\title{Write pngs to temporary file for sending via the API}
\usage{
write_temp_png(data)
}
\arguments{
\item{data}{raw list data as parsed from json}
}
\description{
Write pngs to temporary file for sending via the API
}
