% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{app_ui}
\alias{app_ui}
\title{app_ui}
\usage{
app_ui(req, nquestions)
}
\arguments{
\item{req}{Internal parameter for \code{{shiny}}.
DO NOT REMOVE.}

\item{nquestions}{integer defining the number of questions for the model creation step,
typically generated from a call to \code{read_csv("setup_questions.csv") \%>\% nrow()}}
}
\description{
Defines the user interface for the DIAGRAM application
}
