\name{openplotmat}
\alias{openplotmat}
\title{Creates an empty plot used for diagram plotting}
\description{creates a plotting region, bounded by [0,1] 
    without axes, labels, titles}
\usage{openplotmat(asp=NA,...)}
\arguments{
  \item{asp }{the y/x aspect ratio}
  \item{... }{arguments passed to \link[shape]{emptyplot}}
}
\author{Karline Soetaert <k.soetaert@nioo.knaw.nl>}
\seealso{\code{\link[shape]{ emptyplot}} from package shape}
\keyword{hplot}

