% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{score_model}
\alias{score_model}
\title{Score model}
\usage{
score_model(model, responses)
}
\arguments{
\item{model}{a model of class bn.fit}

\item{responses}{a \code{tibble(node, response)} where node is the
name of the node in the network and response is the user value(s) from the
associated question}
}
\value{
a named list of two scores
}
\description{
This function will take a set of user values as a tibble and return the
utility scores from the model given those inputs
}
