% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_validate.R
\name{validate_json}
\alias{validate_json}
\title{Validate json}
\usage{
validate_json(req)
}
\arguments{
\item{req}{a rook request object}
}
\description{
Validate whether the json data posted fits with
the restrictions of expected data structure for DiAGRAM
}
